Jeeves class, main method: Update to be responsive to user input
The program only starts and ends instantly.

There is a need for it to be able to recognize user input and respond accordingly.

Let's add the ability to read and recognize user input.
Specifically, Jeeves can recognize the 'bye' command and terminate appropriately.
All other commands will simply be echoed back.