Implements ToDos, Deadlines and Events, as well as input/output (i.e. commands and printing to Kniaz's command line) for said tasks, and refactors Task printing slightly.
Feature expansion to hit Level 4.

ToDos are Tasks without dates/times, Deadlines are Tasks with a singular due date/times, Events are Tasks with a start/stop.

Also refactors Task to be an abstract class for ToDo to inherit from.

This is a slight shift of paradigm, as a Task would be something you need to do, but alone contains insufficient information.

ToDos lack time as it's just something the user has to do -- time isn't relevant here

A deadline naturally has a due date.

Events in Kniaz are defined by start-stop in addition to the name, which is a natural abstraction of an event (Kniaz doesn't care about place (yet?))

Lastly, string representation to output of Tasks in Kniaz's output has been shifted from toString() to toPrintString(), to diffrentiate between input meant SOLELY for the user (which is toPrintString())