Modified expected file to match the expected output.
User input is entered after a print(...) line instead of a println(...) line, and does not have a newline inside of it. As a result, the test will have divider lines overlapping with the user console. This is not very desirable behavior in the use case where users redirect their input into the program, but as this app is primarily a console-based program, we will not fix this yet.