Allow saving and loading tasks from hard drive
Tasks are currently saved in an ArrayList, which disappears when the
program restarts.

Saving the tasks such that they are persistent will make the program
more useful and prevent data from being lost when the program quits.

Let's store the tasks in a text file in the data directory and update it
accordingly when the task list changes, and obtain tasks from it when
the program starts.