Add isOnDate method to all Task classes
We want to be able to pass in a LocalDateTime to a Task
and be able to tell if the Task has a Date that falls
within the queried Date.

This is needed to support the date query function to
find what deadlines and events fall within the queried date.

Let's apply polymorphism such that all subclasses of Task
override the comparison method in Task.