Save tasks when list changes and load data on start.
Currently, tasks added or updated are not saved when the program is terminated.

Tasks must be saved as users may require the list of tasks the next time they use the program.

Let's save accumulated tasks in a duke.txt file automatically whenever the task list changes and load the task list from the file when the chatbot starts up.

Having a specific file to store the task list allows changes to be saved and loaded easily.