Refactor codebase to allow Storage to become nonstatic
Initially the data text file was hardcoded in many locations.
We want to encapsulate that into the Storage class.

Let's make Storage nonstatic, reducing the need to pass
around dataFilePathname. Let's also revise storage to
create a new ./data/dot.txt file if it doesn't exist.

This also reduces the amount of parameters, making the code
cleaner in general.

It is good to note that running the source code Dot.java
will do it for the project root ip. For text-ui-test,
it is actually executed in the directory text-ui-test.
For JUnit, its runs it in the project root.