Understand Deadlines.
Deadlines are treated as simply strings.

We need to have some form of checking the string input to ensure that a proper date and time is inputted.

Let's verify that the date and time inputted by the user is valid and store it in the Deadline object as a LocalDateTime object.

By doing so, we can be sure that the user has inputted a valid deadline and code is more organised.