Change the date format to java LocalDate
Previously the date format is in String, users can just type anything.
There was no input checking/ or was difficult. Changing the date format
makes the date more standardised.

The input String for event /from, /to and deadline /by,
now only take in the format of "yyyy-mm-dd".
Other format will be considered as invalid input.
The valid string will be parsed as java LocalDateObject.