Update code to allow for save
Bot refreshers with each load.

Users will not be able to use the bot effectively as data has to be
re-entered every load.

Let's save the user's input when assigning tasks to the bot.

Bot will now be able to reference the saved inputs and load the user's
saved data.