Structure the code using more OOP.
Most of the code is cluttered in the Duke.java file.

This makes the card hard to read, and makes it harder to change when the code is not segmented.

Let's edit the code and use more OOP!

Incorporating more OOP makes the code easier to read and more scalable as related fields and methods are grouped together.