Encapsulate files in a duke package.
All files are currently in the src/main/java directory.

The files must be encapsulated as it is easier to determine which files are related to which.

Let's place all files in a package and include them in that package.

By putting all files under a package, it encapsulates the Duke chatbot and its functionalities.