Update bot to understand date and time
There is no fixed format to the input for by, from and to. Bot is able
to accept any input and store them as a valid date and time.

Let's introduce LocalDateTime to the bot to enforce a specific input
format for the date and time input.

This minimises the random behaviour of the bot as any incorrect inputs
will be rejected by the bot.