Change the design to OOP
Creating a lot of java files:
Parser: deals with making sense of the user command.

Storage: deals with loading tasks from the file and
saving tasks in the file.

Command: the action to be execute.

Ui: deals with interactions with the user.

TaskList: contains the task list e.g.,
it has operations to add/delete tasks in the list.