Update code to be more OOP
The code works as intended.

Code can be further extracted to improve OOP.

Extracting the common behavior into a super class allows us to use
polymorphism when dealing with command attributes.

Let's extract and group similar commands in to various classes that
extends the parent class Command.

This improves readability and allows for inheritance as well.