Apply Coding Conventions
Currently, the code does not follow a standard convention.

Code can be hard to read if coding styles are not standardised.

Let's use the SE-EDU Java Coding Standard to standardise the code!

The SE-EDU Java Coding Standard provides a clear and concise coding standard which ensures the code is structured and readable.