Add FindCommand class
This commit adds the FindCommand class, which allows users to find tasks with a specified keyword.

The FindCommand class is a subclass of the Command class. It has a constructor that takes the keyword as a parameter. The execute() method of the FindCommand class calls the getTasksByKeyword() method of the TaskList class to get the tasks that match the keyword. The showTaskList() method of the Ui class is then called to display the list of tasks to the user.