Implement LocalDateTime for Events
The start and end dates inputted into events are not checked whether or not they are actual dates.

This can result in the start and end dates being words instead of dates.

Let's use LocalDateTime for Events as well.

In this way, the start and end dates can be determined efficiently and and effectively.