Add more classes to support OOP
Previously all logic is done in the Duke class. More OOP style was needed.

This commit adds the following classes:
- Command
- DukeException
- Parser
- Storage
- TaskList
- Ui

The respective logic is moved to the respective classes, to support OOP.
The Duke class is now a class that handles the main logic of the program.

However, Command could still be broken down into more classes, and there
may also be other classes that could be added to better support OOP.