Add Command class
TrackerBot currently handles each individual Command case using
a switch case and an enum.

As we switch over to a more rigid OOP structure, it becomes unclear
where this handling should be made.

By extracting the Command class, we can handle it all in the
Command class, and call command.execute().