Add Gradle support for Chat Bot
Previous version of ChatBot did not have support for Gradle.

Support for gradle is required as we ramp up automated unit testing.
Gradle support is therefore added, with JUnit tests coming in the next commit.
Gradle is used for its ease of use and compatibility with JUnit.

Fixed bugs regarding error messages not printing out.