Add JUnit tests into IP
No automated unit testing was implemented in this project. This led to time inefficiency as I had to
manually test each component one by one.

Therefore, JUnit tests are being implemented to test the various components in this project. Currently,
I am implementing only tests for TaskList and Parser. More will be implemented as required.

Fixed a bug regarding InvalidDescriptionException not being thrown when an empty description is passed through.