Update Task and all other related classes for GUI
The move from text UI to GUI requires us to define
handlers that will update the GUI, instead of directly printing.

It is essential to refactor the whole codebase to follow
the GUI change.

Let's take all methods that print to the previous text UI
and turn them to methods that can receive a Consumer
which will update the GUI. We also make a handler capable
of handling errors, that is show the error in the GUI.

Passing GUI handlers around is easier than passing the
to-be-displayed output.