Flip the if statemement to keep indentation level low
There are several points in the code where indentation gets a bit
unwieldy. This is because the code is wrapped inside an if statement.

Flipping the clause of the `if` to a guard clause design pattern will
allow the indentation level to stay low.

This will improve maintainability and readability for the future.