Add extra features in GUI
The interface is lacking in visual information. It is difficult to keep
track of the conversation given the similar looking texts.

Dividing the interaction into the user's input and the program's output
makes the communication more clear and convenient.

To make the interaction more pleasing, let's add graphical user
interface that simulates communicating with another person.