Implement functioning GUI
Previously the GUI implementation was barebones and the UI class was handling messages displayed in the text based UI.

Changed the command classes to generate the messages required by the UI, rendering the UI class obselete. UI class has been deleted and a more functional GUI was implemented that managed to output the messages previously used by Puke.