Add assertions to code
Currently, the code has no assertions and the lack of hard-to-find bugs
is purely contingent on good usage of the code, which is hard to achieve
as the codebase grows.

Adding assertions ensures that functions are being used in the way they
are intended.

Let's add assertions to our code. Also because we are required to do so.