Refactor Command and Task classes into packages
Currently, the Command and Tasks classes are written in one file each
with their subclasses.

This is bad because big files are hard to maintain.

Breaking the files up into smaller files will make it easier to
maintain.

Let's break up the Command and Task class files into smaller components
and put them into packages.

We use packages so because of the natural grouping of tasks and
commands.