Implement code quality
load() in Storage.java: method too long

The load() method is too long and it exceeds 30 lines of code.

Abstracting out repetitive parts of the method makes it more readable for others working on the same code base.

Defining a method to add stored tasks, addStoredTask(), shortens the load() method and makes it more understandable.