Add assertions to some classes
There lacks a way to specify assumptions in the main code.
For instance, in Validation::getArgsIfValidDeadlineFormat,
we made the assumption that a regex is of a certain form
previously in comments.

Let's utilise assertions to ensure that the input is indeed
in the assume formed at that point in the code. We also assert
several conditions at other places in the codebase.

Assertion failures can indicate potential bugs in the code due
to an invalid assumption, which can then be corrected.