Add assertions to ensure correctness of code
There are some methods that amend the underlying implementation
of the TaskList, and correctness (length of the list after amendments)
needs to be checked to ensure that invariants are not violated.

Adding assertions to the code improves code quality and correctness.

In an effort to ensure correctness, let's include some assertions at
crucial points of the code where correctness is important. Doing so
will improve code correctness.