Add guard clauses to make happy paths prominent
The main code, should no exception be thrown, is nested
within conditionals, on the same level as exception throws.

Let's implement some guard clauses, in order to allow the
happy path to be at the top layer, improving readability.

Having a prominent happy paths is more desirable, not only
due to its readability, but in the case we want to add on
to this happy path in the future.