Adding assertions to UI elements and task index
The code has some assumptions about the UI elements
being present and task indexes being positive at runtime.

However this might not be the case, so missing UI elements and
negative task indexes can lead to unexpected behaviours.

Adding assertions can help us to catch these missing UI elements
and negative indexes and stop the execution.

Using assertions is preferable as it is easy to implement, easy to
disable for production while ensuring the code functions as expected.

Refer to this on the use of assertions:
https://nus-cs2103-ay2324s1.github.io/website/se-book-adapted/chapters/errorHandling.html#assertions