Add friendlier syntax and remove parameter order
Currently parameters must be specified in a specific order and
the root command can be long to type.

This makes it troublesome of the users to remember the order of parameters
as well as the full word of the command.

We now allow for parameters to be specified in any order while still
having proper input validation for existence and typing of it.
Commands are also given shortened forms as shortcut (e.g. delete: rm).

This way we allow for greater flexibility and make the CLI more intuitive to
the user.