Only exceptions are being used to handle errors.
Assertions can help verify certain assumptions in the code, and can
complement exceptions to handle errors in the code.

Inserting assertions into help verify that no invalid inputs are
being passed into Commands and that a response will be shown to
the user.

Assertions and exceptions complement each other to handle errors.