Add assertions
We are not able to pass an instance of Shiba to MainWindow, and vice
versa, so we have singletons for both of these classes instead. Hence,
we check that the singleton instance does not already exist when
creating a new instance.

We also assert that a file exists after creating it with no errors.