Add missing JavaDocs
Due to major changes in code structure in `Level-10`, many new public
methods were introduced, and some methods function very differently
from before.

This necessitates the need for an update in the JavaDocs so that users
and reviewers can better understand the intentions of the methods,
improving their understanding of the flow structure.

As a further step towards improving code structure, it is proposed that
we create a new Command interface and model all Commands as classes to
implement the interface. This adds a further layer of abstraction and
improves adaptability of code in the long run.