Implement assertion statements
There are no assertion statements throughout the program.

Assertion statements are important in ensuring that certain key values
are as expected to ensure that Duke runs properly.

Let's write assertion statements in the Event and Deadline classes so
that we can ensure that the important variables get initialised and stored
properly.