Add assertions for important assumptions in program
There are some important assumptions made in the program made in
the program that should hold. For example, arguments should not
be an empty string, etc.

This allows for early catching of bugs in the program.

Java Assertions can be used to assert a certain value in the program
at a certain point in the code. Hence, adding assertions to assert
a certain value can target this problem.