Improve code quality by refactoring methods
Long method with too many lines could cause future developers reading
the code from quickly understanding what the method is doing. It could
also be a sign that the method is not following the SLAP principle.

There could also be high level of nestings leading to the arrowhead
style code - making it harder for people to read the code.

By extracting certain parts of the method into their own method can
reduce the number of levels of abstraction in a single method.