Update FindCommand and Ui methods Using Streams
FindCommand and Ui methods uses traditional loops to iterate
through the list of tasks and manually build a message to display
to the user.

The code is less concise, more error-prone, and affects readability.

Using Streams abstracts away the low-level details of iteration and
string manipulation. It allows for a more declarative and functional
approach to processing the task list, resulting in cleaner and
more maintainable code.

Let's,
* Update FindCommand to use Streams for task list filtering.
* Update showList to use Streams for task list formatting.
* Update showTasksOnDate to use Streams for task list formatting.
* Update showTasksForToday to use Streams for task list formatting.
* Update showMatchingTask to use Streams for task list formatting.