Add priority command
Different tasks have different priorities that signify its urgency, but it cannot be included currently.

Tasks can now have a priority level assigned to them from 0 to 3. The default priority level of a task is set to 0.

To change the priority level of a task, use the priority command "priority id level",
where id is the index of the task and level is the priority level to set to.