Add welcome message for new users
Currently there is no difference in the greeting message for returning or new users.
Hence it is difficult for a user to know if he or she has used the chatbot before.

With the addition of a welcome message of "Welcome new user!!!", the user
will definitely know if this is a new account with new data.

It can also be used to verify that all previous data has been cleared and
the chatbot is reset.