Change the chatbot message when there are no tasks in chatbot.
Previously, regardless of how many tasks are in the chatbot, when the user
sends the list command, it gives the same message "Here are the tasks in your list:",
followed by the list of tasks. However, when there are no tasks, the message seems to be
strange where it describes an empty list.

The message is now changed to show "You have no tasks currently." when there are no
tasks in the chatbot. This is now more natural as a dialog and is straight and concise.