Add assertions to the codebase
This is needed because we want to avoid regressions to the software as
more features are being added.

Add assertions to the tasklist to check that the index is valid before
we mark, unmark or delete a task and also to the todo, deadline and
event tasks to check if there are the right number of elements in the
arrays after splitting before further formatting can be done.

This ensures more robustness in the codebase.