Add basic null value assertions
There are several parts of the program flow that will be
affected if a null value is assigned instead of its intended
type of value.

Adding assertions to check for potential null values can help
minimise or prevent bugs.

Assert statements has been added with specific messages for
ease of identifying.