Add reminder feature to the bot
Within 7 days of a deadline, the bot will add it to its reminders,
showing the information of the tasks when it starts running and
when the user wants to see the list of tasks

Improve code quality for Storage class and the various Task classes

Split long methods into smaller methods to maintain readability. Doing
so will make future editing easier.