Refactor code to be more readable
Several new methods are extracted from long methods such as
parseCommand from the parser class to help with readability.

parseCommand itself is modified along with several other methods
to return a String directly rather than rely on the Ui class.

This keeps the Ui class methods restricted to only functions
that involve adding and deleting tasks which simplifies the code.