Implement mark multiple and unmark multiple command
Commands to executes mass operations are widely seen in many text
based operations. The ability to execute mass operations can
increase the efficiency and improve the experience of the user.

The "markm" command and "unmarkm" command allows users to input
multiple evenly spaced integer numbers to mark the corresponding
tasks as done.