Refactor task list handling
TaskList previously violated Single Responsibility Principle since
it was handling the parsing of data from the storage. The parsing
and deserialization of the data has now been shifted to a new class
TaskDeserializer.

TaskDeserializer takes in a line of serialized Task and returns the
deserialized Task. It also performs validation to check if the
data file is corrupted.

Storage now provide methods to directly save and fetch
TaskList. TaskDeserializer is used by Storage to save a TaskList
instance into file. The commands utilizes the new methods to reduce
code duplication.