Add the ability to deal with duplicate items
When the user has duplicated tasks in the duke.txt file, or inputs
duplicated tasks, it will be added into the list of tasks.

This makes it confusing for the users when reading the list
of task as those duplicated tasks are redundant.

When loading the file and inputting new tasks, the code now checks
for duplicated tasks and only add the unique task.
In checking for duplicated tasks, the code compares the task
description, and the dateTime where applicable
without considering the seconds.

Checking for duplicated tasks is preferred for better clarity
and readability for users when viewing the list of tasks.
Ignoring the seconds when checking for duplicated tasks is preferred
as the seconds is insignificant when viewing tasks but provides
more flexibility in checking for duplicated task.

Refer to this to deal with duplicated tasks:
https://nus-cs2103-ay2324s1.github.io/website/se-book-adapted/projectDuke/index.html#extensions-category-c