Start A-CI
Current project is not set up with Continuous Integration.

Continuous Integration and Continuous Delivery enhance efficiency for shipping software.

GitHub Actions will run the workflow (as per the .yml file) every time certain project events are triggered (e.g., when a PR is updated, or the master branch is updated).