Add assert to stop execution of programme given file corruption
There are no remedial action for default statements used under switch
statements

Amended the classes above in the commit: DateTime, Parser, TaskList
In DateTime:
* If we are to create a DateTime from storage and a WrongInputException
is thrown, that means that there is data corruption in the file storage
That is a invalid file format was stored. Thus we use assert to
terminate the program since the file is no longer safe to work with.

In Parser:
* Upon hitting the default case which means after parsing a command,
it did not throw an exception however it is invalid means that the
system has failed to detect an invalid command, the program should
terminate to prevent any further actions

In TaskList:
* We assert to check if the storage data is corrupted, that is if there
is even the minimum amount of data to form a task (Task name, status
and type). Programme should terminate to prevent wrong exception
message for the problem
* Assert in default statement since it is a invalid task however it was
created and stored , data is corrupted, should stop execution of program
* Assert in creating task List because we return null in ParseTask
This results in problematic null pointer exception that
could possibly be caught and given the wrong description.
Program is no longer functioning as it should be, should terminate

The asserts are done as a way to deal with highly unlikely but
theoretically possible situations that could cause multiple exceptions
and code functionality to stop working as intended. The strategy to
recover from this situation is to terminate since clearing data file
should not be called by default