FixedDuration task class: Add support for FixedDuration task type
CR7 did not provide support for tasks with a fixed duration but no
start or end time.

Adding the FixedDuration task class allows users to manage such tasks.

Let's add the FixedDuration class as a subclass of the Task class.

Using inheritance as Fixed Duration tasks are a specific type of task
with similar characteristics and hence suitable as a subclass.