Update assertions
Include assertions in Storage constructor, saveTask method in the Storage class to verify that filePath and taskList are not null.
Include assertion in remove method in the TaskList class to verify that taskNumber is not negative.
Include assertions in markTask, unmarkTask and delete methods in TehO class to verify that taskNumber is not negative .