Complete A-CodeQuality: Improve Code Quality
Currently, there are certain coding practices that are not corrected.
This reduces the readability of the code for other programmers.

To correct this, the code has been refactored to enhance organisation
so that there is less code repetitions and a more coherent flow,
making it easier for other programmers to trace the code.