Complete A-Assertions: Use Assertions
There is currently insufficient checking of inputs and
variables to ensure that assumptions hold. This would be
problematic if the assumptions fail to hold, as it results
in incorrect running of the chatbot.

To rectify this, assert statements should be added to test
for such assumptions. This will signal to programmers in the event the
assumption fails to hold, enabling programmers to rectify the error.