Add an extention B-Reminders
There is currently no way to track whether the due
date of the deadline or event is getting close. Users
may have to manually search up the entire list to
see which is due soon.

This feature will help to filter out tasks
due soon, and remind users of these tasks and their
deadline upon start up and also upon the use of the
'remind' command.