Implement C-Sort
The task list cannot be sorted by their categories.

A sorted task list allows the user to view their tasks in a more
organised manner, making it easier for them to differentiate the tasks
by their categories.

Let's create a TaskCategoryComparator class that implements the
Comparator interface and a SortCategoryCommand class.

Creating a TaskCategoryComparator class allows us to use the sort method
in the Collections class to sort our array list of tasks.